+++ src/rating/index.tsx
@@ -1,24 +1,18 @@
-import React from 'react'
-import cc from 'classcat'
+import styled from 'styled-components'
+import { color, space } from '_utils/branding'
 
-import style from 'rating/style'
-import Stars from 'stars'
+import Rating from './Rating'
 
-interface Rating {
-  className?: Classcat.Class
-  score: number
-  ratings: number
-  children: string
-}
+const StyledRating = styled(Rating)`
+  & {
+    display: flex;
+    align-items: center;
+  }
 
-const Rating = ({ className = '', score = 0, ratings = 0, children = '' }: Rating): JSX.Element => (
-  <div className={cc(['kirk-rating', className])}>
-    <Stars stars={score} />
-    <span>
-      {ratings} {children}
-    </span>
-    <style>{style}</style>
-  </div>
-)
+  & span {
+    color: ${color.secondaryText};
+    margin-left: ${space.s};
+  }
+`
 
-export default Rating
+export default StyledRating
